+++ src/app/model/rdf/sparql/datacube/package.scala
@@ -1,6 +1,6 @@
 package model.rdf.sparql
 
 package object datacube {
-  type MeasuresByKey = Map[String, Option[Float]]
+  type MeasuresByKey = Map[String, Option[Long]]
   type SlicesByKey = Map[String, MeasuresByKey]
 }
